# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentStamp API",
    "version": "2.3.0",
    "description": "AgentStamp provides cryptographic identity certificates (stamps), a searchable public registry, reputation scores, cross-protocol passports, verification badges, and a Wishing Well marketplace for AI agents. Free tier available. Paid endpoints use x402 micropayments (USDC on Base and Solana).",
    "contact": {
      "name": "AgentStamp",
      "url": "https://agentstamp.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://agentstamp.org",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/passport/{walletAddress}": {
      "get": {
        "operationId": "getPassport",
        "summary": "Get full signed passport",
        "description": "Returns a full signed agent passport bundling identity, stamp, reputation, A2A card, and MCP metadata into a single Ed25519-signed verifiable credential.",
        "tags": ["Passport"],
        "parameters": [
          {
            "name": "walletAddress",
            "in": "path",
            "required": true,
            "description": "The wallet address of the agent",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed passport",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassportResponse"
                }
              }
            }
          },
          "404": {
            "description": "No active agent found for this wallet address",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ErrorResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "register": {
                          "type": "string",
                          "example": "https://agentstamp.org"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/passport/{walletAddress}/a2a": {
      "get": {
        "operationId": "getA2ACard",
        "summary": "Get A2A agent card",
        "description": "Returns just the A2A (Agent-to-Agent) agent card portion of the passport, compatible with Google's A2A protocol.",
        "tags": ["Passport"],
        "parameters": [
          {
            "name": "walletAddress",
            "in": "path",
            "required": true,
            "description": "The wallet address of the agent",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "A2A agent card (returned directly, no wrapper)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/A2ACard"
                }
              }
            }
          },
          "404": {
            "description": "No active agent found for this wallet address",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string" }
        }
      },
...
      "ReputationBreakdown": {
        "type": "object",
        "properties": {
          "tier": { "type": "number", "description": "Score from stamp tier (max 30)" },
          "endorsements": { "type": "number", "description": "Score from endorsements (max 30)" },
          "uptime": { "type": "number", "description": "Score from heartbeat uptime (max 20)" },
          "age": { "type": "number", "description": "Score from registration age (max 10)" },
          "wishes": { "type": "number", "description": "Score from wishes granted (max 10)" }
        }
      },
...
      "PassportAgent": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "wallet_address": { "type": "string" },
          "category": { "type": "string" },
          "capabilities": { "type": "array", "items": { "type": "string" } },
          "protocols": { "type": "array", "items": { "type": "string" } },
          "endpoint_url": { "type": "string" },
          "status": { "type": "string" },
          "registered_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" }
        }
      },
      "PassportStamp": {
        "type": ["object", "null"],
        "properties": {
          "id": { "type": "string" },
          "tier": { "type": "string" },
          "issued_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" },
          "valid": { "type": "boolean" },
          "verify_url": { "type": "string" }
        }
      },
      "PassportReputation": {
        "type": ["object", "null"],
        "properties": {
          "score": { "type": "integer", "minimum": 0, "maximum": 100 },
          "label": { "type": "string", "enum": ["new", "emerging", "established", "elite"] },
          "breakdown": { "$ref": "#/components/schemas/ReputationBreakdown" }
        }
      },
      "A2ACard": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "url": { "type": "string" },
          "provider": {
            "type": "object",
            "properties": {
              "organization": { "type": "string" },
              "url": { "type": "string" }
            }
          },
          "version": { "type": "string" },
          "capabilities": {
            "type": "object",
            "properties": {
              "streaming": { "type": "boolean" },
              "pushNotifications": { "type": "boolean" },
              "stateTransitionHistory": { "type": "boolean" }
            }
          },
          "authentication": {
            "type": "object",
            "properties": {
              "schemes": {
                "type": "array",
                "items": { "type": "string" }
              },
              "credentials": {
                "type": ["object", "null"],
                "properties": {
                  "stamp_id": { "type": "string" },
                  "tier": { "type": "string" },
                  "verify_url": { "type": "string" }
                }
              }
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" },
                "description": { "type": "string" }
              }
            }
          },
          "extensions": {
            "type": "object",
            "properties": {
              "agentstamp": {
                "type": "object",
                "properties": {
                  "reputation_score": { "type": "integer" },
                  "reputation_label": { "type": "string" },
                  "stamp_tier": { "type": "string" },
                  "profile_url": { "type": "string" }
                }
              }
            }
          }
        }
      },
      "PassportMCP": {
        "type": "object",
        "properties": {
          "server_url": { "type": ["string", "null"] },
          "capabilities": { "type": "array", "items": { "type": "string" } },
          "supported_protocols": { "type": "array", "items": { "type": "string" } }
        }
      },
      "PassportVerification": {
        "type": "object",
        "properties": {
          "public_key": { "type": "string" },
          "algorithm": { "type": "string", "example": "Ed25519" },
          "verify_endpoint": { "type": "string", "example": "https://agentstamp.org/.well-known/passport-public-key" }
        }
      },
      "Passport": {
        "type": "object",
        "properties": {
          "version": { "type": "string", "example": "1.0" },
          "type": { "type": "string", "example": "AgentPassport" },
          "issued_at": { "type": "string", "format": "date-time" },
          "issuer": { "type": "string", "example": "https://agentstamp.org" },
          "agent": { "$ref": "#/components/schemas/PassportAgent" },
          "stamp": { "$ref": "#/components/schemas/PassportStamp" },
          "reputation": { "$ref": "#/components/schemas/PassportReputation" },
          "a2a": { "$ref": "#/components/schemas/A2ACard" },
          "mcp": { "$ref": "#/components/schemas/PassportMCP" },
          "verification": { "$ref": "#/components/schemas/PassportVerification" },
          "signature": { "type": "string", "description": "Ed25519 signature of the entire passport document" }
        }
      },
      "PassportResponse": {
        "type": "object",
        "required": ["success", "passport"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "passport": { "$ref": "#/components/schemas/Passport" }
        }
      },
...
    }
  }
}